+++ src/main/java/com/example/entity/BiologicalCase.java
@@ -11,8 +11,8 @@ public class BiologicalCase {
     int id;
     LocalDate date;
     LocalTime time;
-    int longitude;
-    int latitude;
+    double longitude;
+    double latitude;
     String country;
     String province;
     String urban;
